face-recognition-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
insightface==0.7.3
onnxruntime-gpu==1.16.3
onnxruntime==1.16.3
pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiofiles==23.2.1

